--- pkg/infra/tracing/opentelemetry_tracing.go
@@ -1,1 +1,0 @@
-package tracing
